pkg/database: Add api methods to interact with SQL database.
This changes the flow of how to connect with the DB. There's
an interface called `SQLDB` which implements all the API methods
and the `GetSQLDB` returns the mentioned interface.

This procedure is followed to resolve the segmentation fault
issue when just using a global DB variable. (issue to be resolved)

Signed-off-by: Vaibhav <vrongmeal@gmail.com>