Change check ID type to string (#51)
Previously, we were going to store check IDs as integers but this
caused an issue for checks that were going to be configured using
the agent's config file. Changing it to a string, which will have
to be unique, caters to both the audience. In the configuration,
the user can provide a unique sluggish string and from the frontend,
a unique slug can easily be generated.

Resolves #45