influxdb: Fix query for collecting metrics.
Previously the regex to match all the IDs matched any char
and not the complete word, limited by start and end. For
example: if check IDs were "abc" and "def", it matched all
the characters from 'a' to 'f'. This behaviour is fixed by
updating the regex to only select the check IDs that are
provided to us. This match is exact and literal.

Fixes #90

Signed-off-by: Vaibhav <vrongmeal@gmail.com>